requirements.txt
openai~=1.65.3

python-dotenv~=1.0.1

supabase~=2.13.0
